Fix Codacy issues
