+++ KeywordSearch/src/org/sleuthkit/autopsy/keywordsearch/Ingester.java
@@ -152,12 +152,12 @@ class Ingester {
         try (BufferedReader reader = new BufferedReader(sourceReader)) {
             Chunker chunker = new Chunker(reader);
             while (chunker.hasNext()) {
-                Chunk chunk = chunker.next();
-
                 if (context != null && context.fileIngestIsCancelled()) {
                     logger.log(Level.INFO, "File ingest cancelled. Cancelling keyword search indexing of {0}", sourceName);
                     return false;
                 }
+
+                Chunk chunk = chunker.next();
                 Map<String, Object> fields = new HashMap<>(contentFields);
                 String chunkId = Server.getChunkIdString(sourceID, numChunks + 1);
                 fields.put(Server.Schema.ID.toString(), chunkId);
